Bugfix: Avoid infinite loop inside @Atomic method when calling an @Atomic super method
This happened because the callable for the superclass used invokevirtual to call back to the
atomic$method, but due to subclassing rules, the one executed would be the one from the
subclass, as in the below example.

class B extends A {
    @Override @Atomic public void m() {
        System.out.println("Inside B.m()");
        super.m();
    }
}

public class A {
    @Atomic public void m() {
        System.out.println("Inside A.m()");
    }

    public static void main(String[] args) {
        new B().m();
    }
}

After processing, this example would result in an infinite loop.

To solve this, the transformation has been tweaked, and the atomic$method is now made static.
As it is static, the callable can directly call the correct version from the superclass.

(Additionally, this allows some simplifications in ProcessAtomicAnnotations, which is why the diff
is bigger than expected.)